Give party guests the same search dropdown as the quiz setup
